requirements.txt
aiogram>=3.23.0,<4.0.0
aiohttp>=3.11.0
aiosqlite>=0.20.0
python-dotenv>=1.0.1
pydantic>=2.12.0
gigachat